refactor: create app package and move routes to views.py
